Move regex paths to separate file
